finished ensuring both files can be parsed, in theory should work for any file. Minor refactor and code cleaning
